修改为使用Guid作为主键

+++ JxAudio.Core/Service/DirectoryService.cs
@@ -13,7 +13,7 @@ public class DirectoryService
             .OrderBy(x => x.Name)
             .ToListAsync<MusicFolder>(x => new MusicFolder()
             {
-                id = (int)x.Id,
+                id = x.Id,
                 name = x.Name
             }, cancellationToken);
         return new MusicFolders()
